Using datatables and refactorizing code
