--- e2e/features/step_definitions/CommonSteps.js
@@ -1,7 +1,0 @@
-const {Given} = require('cucumber');
-
-Given(/^I log in"$/, () => {
-    // Implement log in as
-});
-
-
