+++ e2e/pages/MainAccountPage.po.js
@@ -2,26 +2,18 @@
 
 const CommonActions = require('../core/ui/CommonActions');
 
-const WAIT_FOR_CLICK = 10000;
-const WAIT_FOR_CREATE = 30000;
+const WAIT_VISIBILITY = 10000;
 
 class account {
 
     constructor() {
-        this.titleBoard = 'input[class="subtle-input"]';
-        this.submit = 'button[type="submit"]';
         this.createBoard = 'a[class="board-tile mod-add"]';
+        this.something = 'a[class="tab__tabLink__3C9rw tab__selected__1gsiC"]';
     }
 
     clickCreateBoard() {
-        browser.waitForExist('a[class="tab__tabLink__3C9rw tab__selected__1gsiC"]', WAIT_FOR_CLICK);
+        browser.waitForVisible(this.something, WAIT_VISIBILITY);
         CommonActions.operate(this.createBoard);
-    }
-
-    createNewBoard() {
-        browser.waitForExist('div[class="board-tile create-board-tile has-photo-background"]', WAIT_FOR_CREATE);
-        CommonActions.setValue(this.titleBoard, 'NewBoard');
-        CommonActions.operate(this.submit);
     }
 }
