--- e2e/pages/MainPage.po.js
@@ -1,14 +1,0 @@
-'use strict';
-
-class MainPage {
-
-    constructor() {
-        this.IdontKnow = '';
-    }
-
-    goUrl(site) {
-        browser.url(site);
-    }
-}
-
-module.exports = MainPage;
